Add to cart and checkout

+++ MeoCamp.Repository/Models/Payment.cs
@@ -15,7 +15,7 @@ public partial class Payment
 
     public string PaymentMethod { get; set; }
 
-    public decimal? Amount { get; set; }
+    public int? Amount { get; set; }
 
     public string PaymentStatus { get; set; }
